% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feeding_rules}
\alias{feeding_rules}
\title{Example feeding rules for \code{traits}}
\format{
\subsection{\code{feeding_rules}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{trait_type_resource}{broader resource trait class - i.e. column name in traits}
\item{trait_resource}{specific resource trait class - i.e. row entry in trait column}
\item{trait_type_consumer}{broader consumer trait class - i.e. column name in traits}
\item{trait_consumer}{specific consumer trait class - i.e. row entry in trait column}
...
}
}

\subsection{\code{feeding_rules}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{trait_type_resource}{broader resource trait class - i.e. column name in traits}
\item{trait_resource}{specific resource trait class - i.e. row entry in trait column}
\item{trait_type_consumer}{broader consumer trait class - i.e. column name in traits}
\item{trait_consumer}{specific consumer trait class - i.e. row entry in trait column}
...
}
}
}
\source{
NA

NA
}
\usage{
feeding_rules

feeding_rules
}
\description{
A mock trait dataset using four trait classes specified in \code{traits}
to infer feeding interactions

A mock trait dataset using four trait classes specified in \code{traits}
to infer feeding interactions
}
\keyword{datasets}
