% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim.R-package.R
\docType{package}
\name{pfim.R-package}
\alias{pfim.R}
\alias{pfim.R-package}
\title{pfim.R: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
