% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traits}
\alias{traits}
\title{Example species traits data to infer feeding rules}
\format{
\subsection{\code{traits}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{species}{species name}
\item{motility}{motility class of species}
\item{habitat}{habitat species found in}
\item{feeding}{trophic level of species}
\item{size}{categorical size classes}
...
}
}
}
\source{
NA
}
\usage{
traits
}
\description{
A mock trait dataset using four trait classes to determine interactions
as specified in \code{feeding_rules}
}
\keyword{datasets}
