% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_edgelist.R
\name{infer_edgelist}
\alias{infer_edgelist}
\title{Infer edgelist using pfim fules blah blah blah}
\usage{
infer_edgelist(
  data,
  cat_combo_list,
  col_taxon = "taxon",
  col_num_size = NULL,
  cat_trait_types = NULL,
  num_size_rule = NULL,
  certainty_req = "all",
  return_full_matrix = FALSE,
  print_dropped_taxa = FALSE,
  hide_printout = FALSE,
  ...
)
}
\arguments{
\item{data}{dataframe object containing species and relevant traits.}

\item{cat_combo_list}{dataframe containing trait rules}

\item{col_taxon}{indicate column containing taxon names.}

\item{col_num_size}{use "size_select" if using a numerical prey-pred size rule indicate column containing numerical size values}

\item{cat_trait_types}{if you only want to use a subset of life habit columns enter column names as vector otherwise}

\item{num_size_rule}{function defining pred-prey size rule, e.g. as long as predator is equal to or larger than prey then an interaction is feasible}

\item{certainty_req}{state whether an interaction must meet all trait rules or the minimum number of rules that must be met}

\item{return_full_matrix}{indicate whether the function should return all taxon pairs with a value indicating how many trait rules were met, or whether the function should just return inferred interactions as an edgelist}

\item{print_dropped_taxa}{print names of any taxa that have been removed}
}
\value{
edgelist showing resource and consumer.
}
\description{
valuable for num_size_rule as an example? \code{function(res_size, con_size) {ifelse(res_size <= con_size, 1, 0)}}
}
