% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim.R-package.R
\docType{package}
\name{pfim-package}
\alias{pfim}
\alias{pfim-package}
\title{pfim: Food Web Construction for paleo communities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R tools for reconstructing paleo food webs from species traits and size rules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beckslab.github.io/pfim/}
  \item \url{https://github.com/BecksLab/pfim}
}

}
\author{
\strong{Maintainer}: Jack Shaw \email{Jack.Shaw@colorado.edu} (\href{https://orcid.org/0000-0002-2725-2334}{ORCID})

Authors:
\itemize{
  \item Tanya Strydom (\href{https://orcid.org/0000-0001-6067-1349}{ORCID})
}

}
\keyword{internal}
