% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfim.R-package.R
\docType{package}
\name{pfwim-package}
\alias{pfwim}
\alias{pfwim-package}
\title{pfwim: Food Web Construction for paleo communities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R tools for reconstructing paleo food webs from species traits and size rules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beckslab.github.io/pfwim/}
  \item \url{https://github.com/BecksLab/pfwim}
}

}
\author{
\strong{Maintainer}: Jack O. Shaw \email{Jack.Shaw@colorado.edu} (\href{https://orcid.org/0000-0002-2725-2334}{ORCID})

Other contributors:
\itemize{
  \item Alexander M. Dunhill (\href{https://orcid.org/0000-0002-8680-9163}{ORCID}) [contributor]
  \item Andrew Beckerman (\href{https://orcid.org/0000-0002-7859-8394}{ORCID}) [contributor]
  \item Baran Karapunar (\href{https://orcid.org/0000-0001-9711-1492}{ORCID}) [contributor]
  \item Tanya Strydom (\href{https://orcid.org/0000-0001-6067-1349}{ORCID}) [contributor]
  \item Jennifer A. Dunne (\href{https://orcid.org/0000-0002-3789-9561}{ORCID}) [contributor]
  \item Pincelli M. Hull (\href{https://orcid.org/0000-0001-8607-4817}{ORCID}) [contributor]
}

}
\keyword{internal}
