% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{powerlaw_prey}
\alias{powerlaw_prey}
\title{Infer edgelist using pfim fules blah blah blah}
\usage{
powerlaw_prey(
  el,
  n_samp = 50,
  func = function(r, M, y) exp(-r/(exp((y - 1) * (log(M)/(y)))))
)
}
\arguments{
\item{el}{edgelist showing resource and consumer. NB col1 = resource, col2 = consumer}

\item{n_samp}{number of samples}

\item{func}{power-law distribution.}
}
\value{
pruned edgelist.
}
\description{
valuable for num_size_rule as an example? \code{function(res_size, con_size) {ifelse(res_size <= con_size, 1, 0)}}
}
