% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{sample_pdf}
\alias{sample_pdf}
\title{Internal powerlaw function used for downsampling metawebs
#'}
\usage{
sample_pdf(
  M = 100,
  y = 2.5,
  func = function(r, M, y) exp(-r/(exp((y - 1) * (log(M)/(y))))),
  n_samp = 100
)
}
\arguments{
\item{M}{richness.}

\item{y}{parameter}

\item{func}{power law.}

\item{n_samp}{number of samples}
}
\value{
TODO
}
\description{
Internal powerlaw function used for downsampling metawebs
#'
}
\examples{
TODO
}
